+++ vars/packer.groovy
@@ -19,6 +19,11 @@ void build(Map config) {
     assert (config.var instanceof Map) : 'The var parameter must be a Map.'
 
     config.var.each() { var, value ->
+      // convert value to json if not string type
+      if (value instanceof List) || (value instanceof Map) {
+        value = writeJSON(json: value, returnText: true)
+      }
+
       cmd += " -var ${var}=${value}"
     }
   }
@@ -243,6 +248,11 @@ void validate(Map config) {
     assert (config.var instanceof Map) : 'The var parameter must be a Map.'
 
     config.var.each() { var, value ->
+      // convert value to json if not string type
+      if (value instanceof List) || (value instanceof Map) {
+        value = writeJSON(json: value, returnText: true)
+      }
+
       cmd += " -var ${var}=${value}"
     }
   }